+++ feature/exercise/src/test/kotlin/io/github/stslex/workeeper/feature/exercise/domain/ExerciseInteractorImplTest.kt
@@ -2,9 +2,12 @@
 package io.github.stslex.workeeper.feature.exercise.domain
 
 import io.github.stslex.workeeper.core.core.images.ImageStorage
+import io.github.stslex.workeeper.core.core.resources.ResourceWrapper
 import io.github.stslex.workeeper.core.exercise.exercise.ExerciseRepository
 import io.github.stslex.workeeper.core.exercise.exercise.model.ExerciseChangeDataModel
+import io.github.stslex.workeeper.core.exercise.session.SessionRepository
 import io.github.stslex.workeeper.core.exercise.tags.TagRepository
+import io.github.stslex.workeeper.core.exercise.training.TrainingRepository
 import io.github.stslex.workeeper.feature.exercise.domain.ExerciseInteractor.ArchiveResult
 import io.github.stslex.workeeper.feature.exercise.domain.ExerciseInteractor.SaveResult
 import io.mockk.coEvery
@@ -23,10 +26,16 @@ internal class ExerciseInteractorImplTest {
     private val exerciseRepository = mockk<ExerciseRepository>(relaxed = true)
     private val tagRepository = mockk<TagRepository>(relaxed = true)
     private val imageStorage = mockk<ImageStorage>(relaxed = true)
+    private val sessionRepository = mockk<SessionRepository>(relaxed = true)
+    private val trainingRepository = mockk<TrainingRepository>(relaxed = true)
+    private val resourceWrapper = mockk<ResourceWrapper>(relaxed = true)
     private val interactor = ExerciseInteractorImpl(
         exerciseRepository = exerciseRepository,
         tagRepository = tagRepository,
         imageStorage = imageStorage,
+        sessionRepository = sessionRepository,
+        trainingRepository = trainingRepository,
+        resourceWrapper = resourceWrapper,
         defaultDispatcher = Dispatchers.Unconfined,
     )